1) Various gcc -Wall fixes; mostly type-punning now resolved by unions.
2) Potential bug fix to scram_merge demo program when using multiple
refs. (On 32-bit systems only?)

3) Added .mft1.00 to the magic number search for 454 indexed SFF
files, in addition tot he old .srt1.00 string. The two seem
compatible.

4) Improved cram_to_bam() handling of aux fields. It identifies the
aux location via bam_aux() now instead of assuming it's part of the
same bam struct. This has no effect for io_lib, but makes the code
work within Samtools.

5) "make check" now runs under a debugging environment if $VALGRIND is
set in the environment. Eg VALGRIND="valgrind --db-attach=yes" make check

+++ trunk/progs/scram_merge.c
@@ -257,13 +257,15 @@ int main(int argc, char **argv) {
 
 	for (j = 0; j < n_input; j++) {
 	    bam_seq_t *b = s[j];
-	    int64_t b_ref, x;
+	    uint64_t b_ref, x;
 	    if (!in[j])
 		continue;
 
 	    b_ref = bam_ref(b) == -1 ? INT_MAX : bam_ref(b);
-	    x = (bam_ref(b)<<33) | (bam_pos(b)<<2) | (bam_strand(b)<<1) |
-		!(bam_flag(b) & BAM_FREAD1);
+	    x = (((uint64_t)bam_ref(b))<<33)
+		| (bam_pos(b)<<2)
+		| (bam_strand(b)<<1)
+		| !(bam_flag(b) & BAM_FREAD1);
 	    if (best_val > x) {
 		best_val = x;
 		best_j = j;
